requirements.txt
annotated-types==0.7.0
dnspython==2.6.1
Inject==4.3.1
pydantic==2.10.4
pydantic-core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
typing-extensions==4.12.2
uuid==1.30
